@@ -3,12 +3,10 @@
     "last 2 Chrome versions"
   ],
   "scripts": {
-    "build-ui": "rollup -c rollup.config.js",
-    "build": "browserify -s epub src/epub/index.js -o bin/book.bin.js",
-    "test": "mocha --delay -t 0 src/epub/test.js",
-    "serve-dev": "nodemon src/server/index.ts"
+    "test": "ts-mocha --delay -t 0 src/epub/test.ts"
   },
   "devDependencies": {
+    "@types/xml2js": "^0.4.8",
     "babel-core": "^6.26.3",
     "babel-preset-env": "^1.7.0",
     "browserify": "^16.2.3",
@@ -18,7 +16,8 @@
     "rollup": "^2.30.0",
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-svelte": "^6.0.1",
-    "xml2js": "^0.4.19",
+    "ts-mocha": "^8.0.0",
+    "xml2js": "^0.4.23",
     "xmldoc": "^1.1.0"
   },
   "dependencies": {
@@ -32,6 +31,6 @@
     "postcss-modules": "^3.2.2",
     "svelte": "^3.29.0",
     "ts-node": "^9.0.0",
-    "typescript": "^4.0.3"
+    "typescript": "^4.2.3"
   }
 }
